+++ spec/spec_helper.rb
@@ -10,9 +10,17 @@ RSpec.configure do |config|
   config.color_enabled = true
   config.filter_run :focus => true
   config.run_all_when_everything_filtered = true
+  config.treat_symbols_as_metadata_keys_with_true_values = true
 
   config.before(:each) do
     @fixture_path = Pathname.new(File.expand_path('../fixtures/', __FILE__))
     @lib_path     = Pathname.new(File.expand_path('../../lib/', __FILE__))
+
+    # Stub all UI methods, so no visible output appears for the UI class
+    ::Guard::UI.stub(:info)
+    ::Guard::UI.stub(:warning)
+    ::Guard::UI.stub(:error)
+    ::Guard::UI.stub(:debug)
+    ::Guard::UI.stub(:deprecation)
   end
 end